requirements/base.txt
# ### Python 3.5.2
# ### http://tecadmin.net/install-python-3-5-on-ubuntu/#
# ## sudo apt-get install python3.5-dev

dj-static==0.0.6
Django==1.11.12
django-localflavor==1.4.1
psycopg2==2.6.2
python-dateutil==2.6.1
python-decouple==3.0
static3==0.7.0
django-test-without-migrations==0.4
djangorestframework==3.11.0
xlrd==1.1.0
xlwt==1.3.0
django-bootstrap3==9.0.0
Pillow==3.4.2
pytz
-e git://github.com/znc-sistemas/django-localbr.git#egg=localbr-dev
django-municipios==1.2.1
	